Refactor class getApps
Move a bunch of the passed in values from the instantiation
of the class into the main() method.  This lets you instantiate
the class once and then call it multiple times with different
values.

Chase those changes in toolkit.py While I was there remove a
bunch of self.x stuff from various methods in toolkit.py.

There's still alot of them left to pull out but to keep the changes
somewhat reasonable per commit.  I only pulled them out of methods
that I had to touch to chase the refactor.

Also, since getApps allows one instatiation -> multiple calls now,
chase that change in toolkit.py and use it.  CPU isn't really an
issue in today's world, but there's no reason to be ineligent if we
can help it.