Add restoreApp to astraSDK.py
This class allows in-place restores of an app from either a
backup or a snapshot.

This is a destructive event in that the current running app is replaced
in place by the contents of the backup or the snapshot.

The method returns True/False, however that only refers to submitting
the restore job itself.  To see if the restore actually succeeds you
must moniter the running state of the application and watch the state
change from "restoring" to either "running" or "failed".