Wire up the PythonSDK restoreApp class to toolkit.py
clone/restore used to be synonyms.

This commit splits those verbs.  A clone creates a new
app in a new namespace from a backup.

A restore destructively overwrites an app in place with a
backup or snapshot of that app.

toolkit.py restore <appID> --backupID OR --snapshotID

If submission of the restore job succeeds then the toolkit.py spins
on the app state (as seen in list apps) waiting for it to go
from "restoring" to either "running" or "failed"