Make toolkit.py list apps line up with the GUI
- No flags shows the apps in the managed tab in the GUI
- -u shows the apps in the unmanaged tab in the GUI
- -i shows the apps in the ignored tab in the GUI

This change addresses https://github.com/NetApp/netapp-astra-toolkits/issues/7

Note that -u and -i are mutually exclusive.  The toolkit enforces this.

When calling getApps() from the Python SDK directly it is possible
to try and list apps that are managed AND ignored.  Since this is
an impossible combination that would always list no apps the SDK
silently changes this to unmanaged AND ignored.

The python SDK uses a parameter called "discovered" to differentiate
between managed and unmanaged.  It would be far better to call this
parameter "managed" however that would involve modifying the callers,
so in the interest of API stability the old name will remain.

Remember that naming things is one of the two hard problems in computer
science.  The other problems being counting things and cache
invalidation.